+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [Unreleased]
 
+### Changed
+
+- Collect matches from all detectors instead of stopping at first match (#16)
+
 ## [0.6.0] - 2026-05-08
 
 ### Changed